requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
django-environ==0.11.2
Markdown==3.6
sqlparse==0.4.4
psycopg==3.1.18
nltk==3.8.1
regex==2023.12.25
sqlparse==0.4.4
tqdm==4.66.2
